web/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
cssselect2==0.7.0
distlib==0.3.6
Django==4.1.7
et-xmlfile==1.1.0
face-recognition-models==0.3.0
filelock==3.9.0
html5lib==1.1
idna==3.4
lxml==4.9.2
numpy==1.24.2
openpyxl==3.1.2
oscrypto==1.3.0
pandas==2.0.0
Pillow==9.5.0
platformdirs==3.1.0
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyMySQL==1.0.3
pypdf==3.7.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
Unidecode==1.3.6
uritools==4.0.1
urllib3==1.26.14
virtualenv==20.20.0
webencodings==0.5.1
xhtml2pdf==0.2.9
xlwt==1.3.0
whitenoise==6.4.0
